feat(sim): compile the production connection stack
The simulator ran the real src/FurbleUI.cpp over four substitutes: a 750 ms
timer Control with four of the six states and no connect abort, a one-boolean
Camera, a process-local CameraList and a two-event fake Scan. The three
2026-08-28 hardware failures all live in the code those substitutes replaced,
so none of them could be reproduced.

Build the production Control, Camera, CameraList, Scan, Device and every
vendor class into the simulator over the shared MockNimBLE and the virtual
camera peers, and delete the substitutes. Real CameraList now persists through
lib/preferences on the simulator preferences file, which closes the saved
camera gap as well.

A new virtual radio task advertises the registered peers to the production
Scan on the plan 158 scheduler gate and owns the discovery timer the
controller owns on hardware, so advertisement matching runs through the real
CameraList::match path. Scenario seeds pick the peer topology; transport
faults replace the simulator side link drop.

Narrow FURBLE_SIM seams retained: the scan start responsiveness probe and the
scan end callback counter move into production Scan as observability, and the
control task counts the camera commands it dispatches so a scenario can still
assert the shutter path fired.

+++ lib/furble/CanonEOSSmart.cpp
@@ -215,11 +215,14 @@ void CanonEOSSmart::updateGeoData(const gps_t &gps, const timesync_t &timesync) 
     time_t timestamp = mktime(&time_str);
     canon_geo_t geo = {
         .header = 0x04,
-        .latitude_direction = gps.latitude < 0.0 ? 'S' : 'N',
+        // The direction fields are uint8_t. Cast the character literals so the
+        // braced initializer is not a narrowing conversion, which clang rejects
+        // outright and gcc only warns about.
+        .latitude_direction = static_cast<uint8_t>(gps.latitude < 0.0 ? 'S' : 'N'),
         .latitude = static_cast<float>(std::abs(gps.latitude)),
-        .longitude_direction = gps.longitude < 0.0 ? 'W' : 'E',
+        .longitude_direction = static_cast<uint8_t>(gps.longitude < 0.0 ? 'W' : 'E'),
         .longitude = static_cast<float>(std::abs(gps.longitude)),
-        .elevation_sign = gps.altitude < 0.0 ? '-' : '+',
+        .elevation_sign = static_cast<uint8_t>(gps.altitude < 0.0 ? '-' : '+'),
         .elevation = static_cast<float>(std::abs(gps.altitude)),
         .timestamp = static_cast<uint32_t>(timestamp),
     };
